Fixed possible problem where a stale lockfile will keep the filesession cleanup from running.
A lockfile is used to ensure that only one process can run the cleanup code
at any given time. A situation may arise where the process running the cleanup
segfaults (ie same process, but different thread) such that the lockfile is not
deleted. If this should happen subsequent attempts to run the cleanup will
fail. To avoid this possibility a test has been added to check the age of the
lockfile. If it is more than 3600 seconds old it will be considered stale
and deleted. The next time filesession cleanup runs it should succeed.


git-svn-id: https://svn.apache.org/repos/asf/httpd/mod_python/trunk@327728 13f79535-47bb-0310-9956-ffa450edef68